Bundler: Modify Bundler manager to avoid download of source codes
Call Bundler inside a ruby script that retrieves the metadata of
each dependency, including transitive ones, without downloading
and installing them. Build analyzer result based on this metadata
and adjust the logic accordingly. Keep the 'bundler_dependencies.rb' script
since it is still useful to parse top-level dependencies and scopes
as a starting point for the dependency tree in the analyzer result.

Signed-off-by: Korbinian Singhammer <external.Korbinian.Singhammer2@bosch.io>