StaticHtmlReporter: Add Software Heritage links to licenses
For an enhanced the user experience, link license names to the respective
license texts. To avoid the use of different sources for SPDX license texts
and texts for LicenseRefs, leverage only Software Heritage for linking to
file content.

A previous idea was to link to the ORT repository itself, but that would
not work for forks of ORT.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>
Signed-off-by: Onur Demirci <onur.demirci@bosch.io>