DependencyNode: Add a getStableReference() function
The function allows querying a reference to a DependencyNode which
remains stable through an iteration over dependencies.

For the Scope-based data structures, this is not yet needed, but will
become important when implementing the DependencyNode interface for
the dependency graph structures. These data structures cannot be
directly mapped to the properties defined by the interface. The
implementation is therefore going to use some tricks with re-using
instances to avoid the creation of a large number of wrapper objects.
This, however, has the consequence, that a normal reference to a
DependencyNode may change its properties.

Extend the documentation to point this out.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>