ProjectAnalyzerResult: Move collectPackagesByScope() to Pub
The function is used exclusively by the Pub package manager. As it
accesses the Project.scopes property, it should not be available at a
central location.

For Pub, it is for now okay to make use of the scopes API. In this
context, this is guaranteed to work. When the package manager is
reworked to use the dependency graph format natively, this usage of
the legacy API will change.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>