OrtResult: Remove a withScopesResolved() call acting as safety net
Due to the way the single ORT commands currently load result files,
this call - executed when constructing the structure of projects - is
not needed, as the transformation to the dependency tree format is
manually performed.

The call, however, blocks the transition to the dependency graph
format as it effectively renders it impossible to access projects from
an OrtResult in the dependency graph format. Therefore, remove it, as
this does not have any impact on the proper functioning of ORT
commands.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>