AnalyzerResultBuilder: Create a dependency graph on building a result
In future, no conversion of the dependency representation is going to
be done on deserialization of ORT results, as the DependencyNavigator
of the result can handle all cases. It is, therefore, also not
necessary to do a conversion to the dependency graph format each time
a result is written.

So, do the conversion to the dependency graph format once when the
result is initially created. This is a more straight-forward mechanism
than using a custom serializer. Another advantage is that it does not
cause unexpected changes on old result files when they are read and
written out again.

As the projects created by the analyzer now use the dependency graph
format per default, also change the dummy project created by
PackageManager in case an exception occurs in resolveDependencies().

Adapt an expected test result slightly, because empty scopes are no
longer output.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>